update plan,transfer balance with password

+++ server/routes/hotspot_plan/hotspot_plan.js
@@ -6,11 +6,12 @@ const bcrypt = require("bcrypt");
 const authorization = require("../../middleware/authorization");
 const validHotspot = require("../../middleware/valid_hot_planInfo");
 
+// const confirmPass = require("../intergrate_Selendra/payment");
 const Payment = require("../intergrate_Selendra/payment");
 
 // Create Account User
 
-router.post("/set-plan", authorization, async (req, res) => {
+router.post("/set-plan", authorization, validHotspot, async (req, res) => {
   try {
     //1. destructure the req.body(username,password)
     // for attributeMD5 & op it is default from database
@@ -34,7 +35,7 @@ router.post("/set-plan", authorization, async (req, res) => {
     } else if (val === 365) {
       optName = "365";
     } else {
-      return res.status(401).json({ message: "Please choose!" });
+      return res.status(400).json({ message: "Please choose!" });
     }
 
     const setPlanAlready = await pool.query(
@@ -42,20 +43,18 @@ router.post("/set-plan", authorization, async (req, res) => {
       [req.user]
     );
     if (setPlanAlready.rows.length !== 0) {
-      return res.status(401).json({ message: "You had already set plan!" });
+      return res.status(400).json({ message: "You had already set plan!" });
     }
 
     ////            check password
-    const pass = await pool.query("select * from useraccount WHERE id = $1", [
-      req.user,
-    ]);
-    // compare password
-    const validPassword = await bcrypt.compare(password, pass.rows[0].password);
-    if (!validPassword) {
-      return res.status(401).json({ message: "Incorect Password!" });
+    const confirm = await Payment.confirm_pass(req, password);
+    if (!confirm) {
+      return res.status(401).json({ message: "Incorrect Password!" });
     }
-    var paid = await Payment.payment(req, asset, value, memo);
-    
+
+    ///////////// check balance with payment /////////////////////////
+    const paid = await Payment.payment(req, asset, value, memo);
+
     if (paid[0] === 200) {
       // 2. enter the user inside database
       await pool.query(
@@ -99,74 +98,37 @@ router.post("/set-plan", authorization, async (req, res) => {
   }
 });
 
-router.put("/reset-plan", validHotspot, async (req, res) => {
+router.delete("/cancel-plan", authorization, async (req, res) => {
   try {
-    //1. destructure the req.body(username,password)
-    // for attributeMD5 & op it is default from database
-    const { username, simultaneous, value } = req.body;
-    const op = ":=";
-    const attributeMD5 = "MD5-Password";
-    const priority = "1";
-    const attributeSim = "Simultaneous-Use";
-    const attributeExp = "Expiration";
-    var val = parseInt(value, 10);
-    var sim = parseInt(simultaneous, 10);
-    var optName;
-
-    //   //  ======---===== For Expiration amount of day =====---======
-    if (val === 30) {
-      optName = "30";
-    } else if (val === 365) {
-      optName = "365";
-    } else {
-      res.status(401).json({ message: "Please choose!" });
-    }
+    const { password } = req.body;
 
     // // if username isn't exist
-    const user = await pool.query("select * from radcheck where username=$1", [
-      username,
+    const user = await pool.query("select * from radcheck where acc_id = $1", [
+      req.user,
     ]);
+    const username = user.rows[0].username;
     if (user.rows.length === 0) {
-      return res.status(401).json({ message: "Account isn't exist!" });
+      return res.status(400).json({ message: "Account isn't exist!" });
     }
-    // delete all histories form radacc, radgroupcheck, radusergroup
-    await pool.query("DELETE FROM radacct WHERE username = $1", [username]);
-    await pool.query("DELETE FROM radusergroup WHERE username = $1", [
-      username,
-    ]);
-    await pool.query(
-      "DELETE FROM radgroupcheck WHERE groupname like '%' || $1 ||'%'",
-      [username]
-    );
 
-    //  insert into table RAD_GROUP_CHECK
-    const sim_Name = attributeSim + "_" + username + "_" + sim;
-    await pool.query(
-      "insert into radgroupcheck(groupname, attribute, op, value) VALUES($1, $2, $3, $4) RETURNING *",
-      [sim_Name, attributeSim, op, sim]
-    );
+    ////            check password
+    const confirm = await Payment.confirm_pass(req, password);
+    if (!confirm) {
+      return res.status(401).json({ message: "Incorrect Password!" });
+    } else {
+      // delete all histories form radacc, radgroupcheck, radusergroup
+      await pool.query("DELETE FROM radcheck where acc_id = $1", [req.user]);
+      await pool.query("DELETE FROM radacct WHERE username = $1", [username]);
+      await pool.query("DELETE FROM radusergroup WHERE username = $1", [
+        username,
+      ]);
+      await pool.query(
+        "DELETE FROM radgroupcheck WHERE groupname like '%' || $1 ||'%'",
+        [username]
+      );
 
-    exp_Name = attributeExp + "_" + username + "_" + optName;
-
-    //   Format Date
-    var due = moment().add(val, "days").format("YYYY MMM DD");
-    await pool.query(
-      "insert into radgroupcheck(groupname, attribute, op, value) VALUES($1, $2, $3, $4) RETURNING *",
-      [exp_Name, attributeExp, op, due]
-    );
-
-    //  insert into table RAD_USER_GROUP
-    await pool.query(
-      "insert into radusergroup(username, groupname, priority) VALUES($1, $2, $3) RETURNING *",
-      [username, sim_Name, priority]
-    );
-
-    await pool.query(
-      "insert into radusergroup(username, groupname, priority) VALUES($1, $2, $3) RETURNING *",
-      [username, exp_Name, priority]
-    );
-
-    res.status(200).json({ message: "Reset plan successfully." });
+      res.status(200).json({ message: "Cancel plan done." });
+    }
   } catch (err) {
     console.error(err.message);
     res.status(500).json({ message: "Server Error!" });
@@ -246,6 +208,10 @@ router.get("/get-plan", authorization, async (req, res) => {
       "select * from radusergroup WHERE username = $1",
       [user.rows[0].username]
     );
+    const timeLeft = await await pool.query(
+      "select * from radgroupcheck WHERE groupname = $1",
+      [detail.rows[1].groupname]
+    );
 
     if (detail.rows.length === 0) {
       return res.status(401).json({ message: "No plan!" });
@@ -264,12 +230,13 @@ router.get("/get-plan", authorization, async (req, res) => {
     if (balance === 365) {
       balance = "365";
     }
+
     res.status(200).json({
       username: user.rows[0].username,
       balance: balance,
       device: a.slice(n1 + 1, a.length),
       plan: b.slice(n2 + 1, b.length),
-      time_left: detail.rows[1].value,
+      time_left: timeLeft.rows[0].value,
     });
   } catch (err) {
     console.error(err.message);
